Refactor board page and db file
There were a lot of features that we haven't implement yet but
had the functions that could perform the respected action
cluttering the db file and so I had to remove them, and will only add them backe when
I'm ready to add any of those features.

I also tried to centralise the type definiitions in the types directory so that they are centralized.