Making websocker server entrypoint for Nextjs
I tried to add a custom server for the websocket functionality
previoulsy, but I wasn't able to get it to work for some reason, so I just resorted to
setting up an express server as a separate folder.

But I just came about an article from socket.io on an easy
way to setup websockets in a Next.js project, and I decided to try it out.