Improve email-verify flow
Previoulsy we passed the user to the api route, but that doesn't
read good UX at all, and so I revamped this, and instead redirected
the user to the server componente that then makes the request to verify the token.

This way we have UI to show the user and not a JSON object, as we did previously.