Tried to manage Neon free-tier connection limit issues
Omo... Neon wanted to frustrate me today, with this their limit for the free tier.
I was struglling with failed queries because of the connection limit set by Neon for using the free-ties
I had to find solutions to this problem because the way the board page is setup, any time there is a falied attempt to get the user(query), we immediately
redirect the user to the login page, this had such a bad user experience.

First I had to reduce the number of concurent queries in some of my database query function
like hte ones thta involded fetching for userId and then populating the "suthorName" field to each of the
response for the chatMessages, and noticeMessages.

Then I also has to manage some configurations.
Overall, there seemed to be a significant improvement in the result.